_examples/hello2.go: fix crash, updates #8
Since lazyInit was introduced, the value of exitHandler cannot be used
before lazyInit is run. While lazyInit is triggered in eval, the value
of exitHandler can be passed to eval when Exit/TExit is used in the first
eval-calling statement.

Add a general forceInit function that could be used if other cases like
this are unearthed.