Added (incomplete) modal and modeless dialog egs
modal.go shows how to create and use a create/destroy modal dialog.
This works in terms of modality but see below.

modeless.go shows how to create and use a show/hide modeless dialog.
This *does not work* in terms of modality since I can't hide/show the
dialog; also see below.

I cannot bind to either dialog's X close button which want to be able to
show how to do (in case there is some action to be taken on
close/cancel).

I can't set or get the spinbox's value.