remove all tests that initialize Tcl/Tk, updates #47
It is known that the ebitengine/purego magic is sometimes not entirely
compatible with go test, see for example

    https://github.com/ebitengine/purego/issues/287

The above report is for darwin, but similar failures happen on other
targets, namely on the developer's main machine - linux/amd64.

The builders were thus only testing if the package builds on all
platforms but all tests that depend on actually using Tcl/Tk libs were
disabled. Mistakenly also non-visual tests were disabled.

The kind of crashes the happen during `go test` are not observed when
code using this package is run outside `go test`, unless Tcl/Tk is
called from a goroutine other than the one that initialited the
so/dll/dylib, for example.

Issue 287 is already closed and a new version is released (0.8.2). It
was tested with no difference in the context of this package - the GUI
initializing tests still fail most of the time.

An expert, fluent in both Go runtime, the purego package and knowledge
of what tricks `go tool` is maybe adding to the game, can perhaps figure
out where is the problem rooted.

The hypothesis is that purego, which out of `go test` seems to work
rock-solid, achieves its magic by interacting withe the Go runtime in
ways that are different when the package is built by `go test`.